add kpartx and mount functionnality

+++ source/generate-retropie-image.sh
@@ -17,7 +17,22 @@
 
 ORIGINAL_RETROPIE_IMAGE=
 GAME_IMAGE=
-RESULT_IMAGE="result.img"
+RESULT_IMAGE="out.img"
+MOUNT_PATH="/mnt/loop0p2"
+
+DEBUG=
+
+# OS pre-requisite #######################################################
+
+function install_prerequesites() {
+
+  # if not Debian then exit
+
+  # Kpartx installation
+  [[ -z $(which kpartx) ]] && sudo apt install kpartx -y
+}
+
+# Image file Management #######################################################
 
 function prepare_image() {
   if [[ -z $ORIGINAL_RETROPIE_IMAGE ]]; then
@@ -31,13 +46,40 @@ function prepare_image() {
     exitonerror_nothingmade "Pas d'image de destination"
   fi
 
-  # Prepare target image
-  rm $RESULT_IMAGE
-  cp $ORIGINAL_RETROPIE_IMAGE $RESULT_IMAGE
+  # Duplicate the original image file into destination image file
+
+  if ! [[ -z $DEBUG ]] && [[ -f $ORIGINAL_RETROPIE_IMAGE ]]; then
+    echo "Debug = ON. Le fichier de résultat existe déjà donc pas de copie"
+  else
+    rm $RESULT_IMAGE
+    cp $ORIGINAL_RETROPIE_IMAGE $RESULT_IMAGE
+  fi
+
+  # check if result file is present
+
   if ! [[ -f $RESULT_IMAGE ]]; then
     local errormsg="Impossible de créer le fichier $RESULT_IMAGE"
-    exiterror_nothingmade $errormsg
+    exitonerror_nothingmade $errormsg
   fi
+
+  # End successfully
+  echo "L'image de destination $RESULT_IMAGE a été corectement générée"
+}
+
+function mount_image() {
+
+  KPARTX_COMMAND=$(sudo kpartx -av ${RESULT_IMAGE})
+  echo $KPARTX_COMMAND
+  sudo mkdir -v ${MOUNT_PATH} 2> /dev/null
+  sudo mount /dev/mapper/loop0p2 ${MOUNT_PATH}
+
+}
+
+function umount_image() {
+
+  sudo umount ${MOUNT_PATH}
+  sudo rm -rf ${MOUNT_PATH}
+  sudo kpartx -d ${RESULT_IMAGE}
 }
 
 # Error Management #######################################################
@@ -61,47 +103,54 @@ function usage() {
 
 function get_options() {
 
-  while getopts ":hi:o:" option ;
+  while getopts "i:o:dh" option ;
   do
-    echo "getopts OPTIND=$OPTIND, Option=$option, OPTARG=$OPTARG, OPTERR=$OPTERR"
+    if [[ -z $DEBUG ]]; then
+      echo "getopts OPTIND=$OPTIND, Option=$option, OPTARG=$OPTARG, OPTERR=$OPTERR"
+    fi
     case "$option" in
 
       #Input Retropie Image
       i)
         ORIGINAL_RETROPIE_IMAGE=$OPTARG
-        echo "Le fichier d'éntrée est $ORIGINAL_RETROPIE_IMAGE"
-	shift
-	#exit 0
-	;;
+        echo "Le fichier d'entrée est $ORIGINAL_RETROPIE_IMAGE"
+	      ;;
 
       #Ouput image result
       o)
-	RESULT_IMAGE=$OPTARG
-	echo "Le fichier de résultat est $RESULT_IMAGE"
-	#exit 0
-	;;
+        RESULT_IMAGE=$OPTARG
+	      echo "Le fichier de sortir est $RESULT_IMAGE"
+        ;;
 
       #Missing Arguments
       :)
-	exitonerror_nothingmade "L'option \"$OPTARG\" requiert une argument"
-	;;
+        exitonerror_nothingmade "L'option \"$OPTARG\" requiert une argument"
+        exit 1
+        ;;
+
+      d)
+        echo "Debug mode = ON"
+        DEBUG="ON"
+        ;;
 
       #Invalid Option
       \?)
-	exitonerror_nothingmade "L'option \"$OPTARG\" est invalide"
-	;;
+        exitonerror_nothingmade "L'option \"$OPTARG\" est invalide"
+        exit 1
+        ;;
 
       # Help
       h)
         usage
         # getting the help message from the comments in this source code
         sed '/^#H /!d; s/^#H //' "$0"
-        exit 0
         ;;
     esac
   done
 
-  #shift $((OPTIND-1))
+  shift $((OPTIND-1))
+
+  # get arguments if any
 
 }
 
@@ -109,6 +158,15 @@ function get_options() {
 
 get_options "$@"
 
+install_prerequesites
+
 prepare_image
+
+mount_image
+
+# if not debug then unmount the image file
+if [[ -z $DEBUG ]]; then
+  umount_image
+fi
 
 exit 0
